Update Neural Network
Updated the neural network to have an Alpha=0.005. Before this change, the prediction based on polling from the last day was showing highly unlikely and unrealistic outcomes.

As an example, of some predictions that I was concerned about was:

Pennsylvania and Michigan: 17 point win for democratic candidate.
(These are both swing states in the election, and I expect a close outcome and not a double digit win.)

Another example is Texas: 13 point win for the democratic candidate.
(Texas is a big Republican State. If Texas was to flip to a democratic state, I would expect it to be a narrow victory, not a double digit victory.)

Based on these outcomes I adjusted the code to correct these issues.